feat: #89 마커 저장 기능 구현
- PlayerViewModel에 ModelContext 주입하여 실제 데이터베이스 저장 지원
- DependencyContainer.makePlayerViewModel()에서 ModelContext 자동 전달
- PlayerViewModel+Marker에 SwiftData import 추가
- updateCurrentMusicMarkers()에서 #Predicate 타입 오류 수정
- 마커 변경사항이 앱 재시작 후에도 유지되도록 개선
